put QPPS vals in
