+++ tests/motors/wheelbase.py
@@ -116,11 +116,12 @@ def init(set_PID=True):
 
 
 	if set_PID:
-		M1QPPS = 180000 #12764
-		M2QPPS = 241547 #12764
-		M3QPPS = 180000 #12764
+		M1QPPS = 380268 #180000 #12764
+		M2QPPS = 302420 #241547 #12764
+		M3QPPS = 411356 #180000 #12764
 
 		# PID stuff here?
 		SetVelocityPID(0, 3.991973876953125, 1.9959869384765625, 5.969512939453125, M1QPPS)
 		SetVelocityPID(1, 3.991973876953125, 1.9959869384765625, 5.969512939453125, M2QPPS)
-		SetVelocityPID(2, 0.0152587890625,   0.6103515625,       0.249481201171875, M3QPPS)
+		SetVelocityPID(2, 3.991973876953125, 1.9959869384765625, 5.969512939453125, M3QPPS)
+		# SetVelocityPID(2, 0.0152587890625,   0.6103515625,       0.249481201171875, M3QPPS)
